requirements.txt
pandas==2.2.3
numpy==2.1.2
matplotlib==3.9.2
seaborn==0.13.2
streamlit==1.39.0
scikit-learn==1.5.2
python-dotenv==1.0.1